+++ x/genutil/legacy/v0_38/migrate.go
@@ -7,13 +7,13 @@ import (
 	v036distr "github.com/cosmos/cosmos-sdk/x/distribution/legacy/v0_36"
 	v038distr "github.com/cosmos/cosmos-sdk/x/distribution/legacy/v0_38"
 	v036genaccounts "github.com/cosmos/cosmos-sdk/x/genaccounts/legacy/v0_36"
-	"github.com/cosmos/cosmos-sdk/x/genutil"
+	"github.com/cosmos/cosmos-sdk/x/genutil/types"
 	v036staking "github.com/cosmos/cosmos-sdk/x/staking/legacy/v0_36"
 	v038staking "github.com/cosmos/cosmos-sdk/x/staking/legacy/v0_38"
 )
 
 // Migrate migrates exported state from v0.36/v0.37 to a v0.38 genesis state.
-func Migrate(appState genutil.AppMap) genutil.AppMap {
+func Migrate(appState types.AppMap) types.AppMap {
 	v036Codec := codec.New()
 	codec.RegisterCrypto(v036Codec)
 
@@ -34,9 +34,7 @@ func Migrate(appState genutil.AppMap) genutil.AppMap {
 
 		// Migrate relative source genesis application state and marshal it into
 		// the respective key.
-		appState[v038auth.ModuleName] = v038Codec.MustMarshalJSON(
-			v038auth.Migrate(authGenState, genAccountsGenState),
-		)
+		appState[v038auth.ModuleName] = v038Codec.MustMarshalJSON(v038auth.Migrate(authGenState, genAccountsGenState))
 	}
 
 	// migrate staking state